move tests one level up
